Fix: added null check to file object fallback

+++ lib/Maintenance/Tasks/LogArchiveTask.php
@@ -100,17 +100,18 @@ class LogArchiveTask implements TaskInterface
             foreach ($fileObjectPaths as $objectPath) {
 
                 $filePath = $objectPath['fileobject'];
-                if ($storage->fileExists($filePath)) {
-                    $storage->delete($filePath);
-                } else {
-
-                    // Fallback, if is not found and deleted in the flysystem, tries to delete from local
-                    $fileRealPath = realpath($filePath);
-                    if (str_starts_with(realpath($fileRealPath), PIMCORE_LOG_FILEOBJECT_DIRECTORY)) {
-                        @unlink($fileRealPath);
+                if($filePath !== null) {
+                    if ($storage->fileExists($filePath)) {
+                        $storage->delete($filePath);
+                    } else {
+    
+                        // Fallback, if is not found and deleted in the flysystem, tries to delete from local
+                        $fileRealPath = realpath($filePath);
+                        if (str_starts_with(realpath($fileRealPath), PIMCORE_LOG_FILEOBJECT_DIRECTORY)) {
+                            @unlink($fileRealPath);
+                        }  
                     }
-
-                }
+                 }
 
             }
